Implement WebPack and modules
